refactor: split web console server modules
