+++ src-tauri/src/modules/web_console/commands.rs
@@ -1,274 +1,18 @@
-//! Local browser console for the full Cockpit Tools UI.
-//! This is intentionally separate from `web_report`, which remains the tokened report endpoint.
+//! Command whitelist for the local web console bridge.
+//!
+//! Keep HTTP transport, event polling, and static file serving outside this file.
+//! New browser-exposed Tauri commands should be added to `dispatch_invoke`.
 
 use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use serde_json::{json, Value};
-use std::collections::{HashMap, VecDeque};
-use std::path::{Component, Path, PathBuf};
-use std::sync::{OnceLock, RwLock};
-use tauri::{Emitter, Listener};
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::net::{TcpListener, TcpStream};
-use tokio::time::{sleep, timeout, Duration};
-use url::Url;
 
-use super::config::PORT_RANGE;
+use super::app_handle;
+use super::events::{
+    emit_web_event_from_browser, register_web_event_listener, unregister_web_event_listener,
+};
 
-const DEFAULT_WEB_CONSOLE_PORT: u16 = 18081;
-const MAX_HTTP_REQUEST_BYTES: usize = 2 * 1024 * 1024;
-const REQUEST_READ_TIMEOUT: Duration = Duration::from_secs(8);
-const EVENT_POLL_TIMEOUT: Duration = Duration::from_secs(25);
-const EVENT_POLL_INTERVAL: Duration = Duration::from_millis(150);
-const MAX_EVENT_QUEUE_LEN: usize = 1024;
-const INDEX_HTML: &str = "index.html";
-
-static ACTUAL_WEB_CONSOLE_PORT: OnceLock<RwLock<Option<u16>>> = OnceLock::new();
-static WEB_EVENT_STATE: OnceLock<RwLock<WebEventState>> = OnceLock::new();
-
-#[derive(Debug)]
-struct HttpRequest {
-    method: String,
-    path: String,
-    query: Option<String>,
-    body: Vec<u8>,
-}
-
-#[derive(Debug, Deserialize)]
-struct InvokeRequest {
-    cmd: String,
-    #[serde(default)]
-    args: Value,
-}
-
-#[derive(Debug, Serialize)]
-struct InvokeResponse {
-    ok: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    value: Option<Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    error: Option<Value>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct WebEventMessage {
-    sequence: u64,
-    event: String,
-    payload: Value,
-}
-
-#[derive(Debug, Default)]
-struct WebEventState {
-    next_listener_id: u32,
-    next_sequence: u64,
-    browser_listeners: HashMap<u32, WebBrowserEventListener>,
-    tauri_listeners: HashMap<String, tauri::EventId>,
-    queue: VecDeque<WebEventMessage>,
-}
-
-#[derive(Debug, Clone)]
-struct WebBrowserEventListener {
-    event: String,
-    client_id: String,
-    registered_after_sequence: u64,
-}
-
-fn web_console_port_state() -> &'static RwLock<Option<u16>> {
-    ACTUAL_WEB_CONSOLE_PORT.get_or_init(|| RwLock::new(None))
-}
-
-fn set_actual_port(port: Option<u16>) {
-    if let Ok(mut guard) = web_console_port_state().write() {
-        *guard = port;
-    }
-}
-
-pub fn get_actual_port() -> Option<u16> {
-    web_console_port_state()
-        .read()
-        .ok()
-        .and_then(|guard| *guard)
-}
-
-pub async fn start_server() {
-    let Some(dist_root) = find_frontend_dist() else {
-        set_actual_port(None);
-        super::logger::log_warn("[WebConsole] frontend dist directory not found, skip startup");
-        return;
-    };
-
-    let mut port = DEFAULT_WEB_CONSOLE_PORT;
-    let mut listener = None;
-    for attempt in 0..PORT_RANGE {
-        let addr = format!("127.0.0.1:{}", port);
-        match TcpListener::bind(&addr).await {
-            Ok(bound) => {
-                listener = Some(bound);
-                if attempt > 0 {
-                    super::logger::log_info(&format!(
-                        "[WebConsole] preferred port {} is busy, switched to {}",
-                        DEFAULT_WEB_CONSOLE_PORT, port
-                    ));
-                }
-                break;
-            }
-            Err(err) => {
-                super::logger::log_warn(&format!(
-                    "[WebConsole] failed to bind 127.0.0.1:{}: {}",
-                    port, err
-                ));
-                port = port.saturating_add(1);
-            }
-        }
-    }
-
-    let Some(listener) = listener else {
-        set_actual_port(None);
-        super::logger::log_error("[WebConsole] no available local port");
-        return;
-    };
-
-    set_actual_port(Some(port));
-    super::logger::log_info(&format!(
-        "[WebConsole] serving full UI at http://127.0.0.1:{}/",
-        port
-    ));
-
-    loop {
-        match listener.accept().await {
-            Ok((stream, _)) => {
-                let dist_root = dist_root.clone();
-                tokio::spawn(async move {
-                    if let Err(err) = handle_connection(stream, dist_root).await {
-                        super::logger::log_warn(&format!("[WebConsole] request failed: {}", err));
-                    }
-                });
-            }
-            Err(err) => {
-                super::logger::log_warn(&format!("[WebConsole] accept failed: {}", err));
-            }
-        }
-    }
-}
-
-async fn handle_connection(mut stream: TcpStream, dist_root: PathBuf) -> Result<(), String> {
-    let Some(request) = read_http_request(&mut stream).await? else {
-        return Ok(());
-    };
-
-    if request.method == "OPTIONS" {
-        return write_response(
-            &mut stream,
-            204,
-            "No Content",
-            "text/plain; charset=utf-8",
-            b"",
-        )
-        .await;
-    }
-
-    if request.method == "POST" && request.path == "/__cockpit_web__/invoke" {
-        return handle_invoke_request(&mut stream, &request).await;
-    }
-
-    if request.method == "GET" && request.path == "/__cockpit_web__/events" {
-        return handle_event_poll_request(&mut stream, &request).await;
-    }
-
-    if request.method == "GET" && request.path == "/__cockpit_web__/health" {
-        let body = json!({
-            "ok": true,
-            "port": get_actual_port(),
-            "version": env!("CARGO_PKG_VERSION"),
-        });
-        let body = serde_json::to_vec(&body).map_err(|err| err.to_string())?;
-        return write_response(
-            &mut stream,
-            200,
-            "OK",
-            "application/json; charset=utf-8",
-            &body,
-        )
-        .await;
-    }
-
-    if request.method != "GET" && request.method != "HEAD" {
-        return write_response(
-            &mut stream,
-            405,
-            "Method Not Allowed",
-            "text/plain; charset=utf-8",
-            b"method not allowed",
-        )
-        .await;
-    }
-
-    let file_path = resolve_static_path(&dist_root, &request.path)?;
-    let (file_path, content_type) = if file_path.exists() && file_path.is_file() {
-        let content_type = content_type_for_path(&file_path);
-        (file_path, content_type)
-    } else {
-        (dist_root.join(INDEX_HTML), "text/html; charset=utf-8")
-    };
-
-    let body = tokio::fs::read(&file_path)
-        .await
-        .map_err(|err| format!("read {} failed: {}", file_path.display(), err))?;
-    if request.method == "HEAD" {
-        return write_response(&mut stream, 200, "OK", content_type, b"").await;
-    }
-    write_response(&mut stream, 200, "OK", content_type, &body).await
-}
-
-async fn handle_invoke_request(
-    stream: &mut TcpStream,
-    request: &HttpRequest,
-) -> Result<(), String> {
-    let invoke: InvokeRequest =
-        serde_json::from_slice(&request.body).map_err(|err| format!("invalid JSON: {}", err))?;
-    let response = match dispatch_invoke(&invoke.cmd, &invoke.args).await {
-        Ok(value) => InvokeResponse {
-            ok: true,
-            value: Some(value),
-            error: None,
-        },
-        Err(error) => InvokeResponse {
-            ok: false,
-            value: None,
-            error: Some(Value::String(error)),
-        },
-    };
-    let status = if response.ok { 200 } else { 400 };
-    let body = serde_json::to_vec(&response).map_err(|err| err.to_string())?;
-    write_response(
-        stream,
-        status,
-        if status == 200 { "OK" } else { "Bad Request" },
-        "application/json; charset=utf-8",
-        &body,
-    )
-    .await
-}
-
-async fn handle_event_poll_request(
-    stream: &mut TcpStream,
-    request: &HttpRequest,
-) -> Result<(), String> {
-    let after = query_param(request.query.as_deref(), "after")
-        .and_then(|value| value.parse::<u64>().ok())
-        .unwrap_or(0);
-    let client_id = query_param(request.query.as_deref(), "clientId");
-    let events = wait_for_web_events(client_id.as_deref(), after).await;
-    let body = json!({
-        "events": events,
-        "latestSequence": latest_web_event_sequence(),
-    });
-    let body = serde_json::to_vec(&body).map_err(|err| err.to_string())?;
-    write_response(stream, 200, "OK", "application/json; charset=utf-8", &body).await
-}
-
-async fn dispatch_invoke(cmd: &str, args: &Value) -> Result<Value, String> {
+pub(super) async fn dispatch_invoke(cmd: &str, args: &Value) -> Result<Value, String> {
     match cmd {
         "plugin:app|version" => Ok(Value::String(env!("CARGO_PKG_VERSION").to_string())),
         "plugin:app|name" => Ok(Value::String("Cockpit Tools".to_string())),
@@ -2688,149 +2432,6 @@ async fn dispatch_registered_app_command(cmd: &str, args: &Value) -> Result<Valu
     }
 }
 
-fn web_event_state() -> &'static RwLock<WebEventState> {
-    WEB_EVENT_STATE.get_or_init(|| RwLock::new(WebEventState::default()))
-}
-
-fn register_web_event_listener(event_name: String, client_id: String) -> Result<u32, String> {
-    let mut state = web_event_state()
-        .write()
-        .map_err(|_| "web event state is poisoned".to_string())?;
-
-    state.next_listener_id = state.next_listener_id.saturating_add(1).max(1);
-    let listener_id = state.next_listener_id;
-    let registered_after_sequence = state.next_sequence;
-    state.browser_listeners.insert(
-        listener_id,
-        WebBrowserEventListener {
-            event: event_name.clone(),
-            client_id,
-            registered_after_sequence,
-        },
-    );
-
-    if !state.tauri_listeners.contains_key(&event_name) {
-        let app = app_handle()?;
-        let captured_event = event_name.clone();
-        let tauri_listener_id = app.listen_any(event_name.clone(), move |event| {
-            push_web_event(&captured_event, event.payload());
-        });
-        state.tauri_listeners.insert(event_name, tauri_listener_id);
-    }
-
-    Ok(listener_id)
-}
-
-fn unregister_web_event_listener(listener_id: u32, client_id: Option<&str>) -> Result<(), String> {
-    let tauri_listener_to_remove = {
-        let mut state = web_event_state()
-            .write()
-            .map_err(|_| "web event state is poisoned".to_string())?;
-        let Some(listener) = state.browser_listeners.get(&listener_id) else {
-            return Ok(());
-        };
-        if client_id.is_some_and(|expected| expected != listener.client_id) {
-            return Ok(());
-        }
-        let Some(listener) = state.browser_listeners.remove(&listener_id) else {
-            return Ok(());
-        };
-        let event_name = listener.event;
-
-        let still_used = state
-            .browser_listeners
-            .values()
-            .any(|registered_event| registered_event.event == event_name);
-        if still_used {
-            None
-        } else {
-            state.tauri_listeners.remove(&event_name)
-        }
-    };
-
-    if let Some(tauri_listener_id) = tauri_listener_to_remove {
-        if let Ok(app) = app_handle() {
-            app.unlisten(tauri_listener_id);
-        }
-    }
-
-    Ok(())
-}
-
-fn emit_web_event_from_browser(event_name: String, payload: Option<Value>) -> Result<(), String> {
-    let payload = payload.unwrap_or(Value::Null);
-    app_handle()?
-        .emit(event_name.as_str(), payload)
-        .map_err(|err| err.to_string())
-}
-
-fn push_web_event(event_name: &str, raw_payload: &str) {
-    let payload = serde_json::from_str(raw_payload)
-        .unwrap_or_else(|_| Value::String(raw_payload.to_string()));
-    push_web_event_value(event_name, payload);
-}
-
-fn push_web_event_value(event_name: &str, payload: Value) {
-    let Ok(mut state) = web_event_state().write() else {
-        return;
-    };
-    state.next_sequence = state.next_sequence.saturating_add(1);
-    let sequence = state.next_sequence;
-    state.queue.push_back(WebEventMessage {
-        sequence,
-        event: event_name.to_string(),
-        payload,
-    });
-    while state.queue.len() > MAX_EVENT_QUEUE_LEN {
-        state.queue.pop_front();
-    }
-}
-
-async fn wait_for_web_events(client_id: Option<&str>, after: u64) -> Vec<WebEventMessage> {
-    let started = std::time::Instant::now();
-    loop {
-        let events = collect_web_events(client_id, after);
-        if !events.is_empty() || started.elapsed() >= EVENT_POLL_TIMEOUT {
-            return events;
-        }
-        sleep(EVENT_POLL_INTERVAL).await;
-    }
-}
-
-fn collect_web_events(client_id: Option<&str>, after: u64) -> Vec<WebEventMessage> {
-    web_event_state()
-        .read()
-        .map(|state| {
-            let active_listeners = state
-                .browser_listeners
-                .values()
-                .filter(|listener| client_id.map_or(true, |id| listener.client_id == id))
-                .cloned()
-                .collect::<Vec<_>>();
-
-            state
-                .queue
-                .iter()
-                .filter(|event| {
-                    event.sequence > after
-                        && active_listeners.iter().any(|listener| {
-                            listener.event == event.event
-                                && event.sequence > listener.registered_after_sequence
-                        })
-                })
-                .cloned()
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-fn latest_web_event_sequence() -> u64 {
-    web_event_state()
-        .read()
-        .map(|state| state.next_sequence)
-        .unwrap_or_default()
-}
-
 fn to_value<T: Serialize>(result: Result<T, String>) -> Result<Value, String> {
     serde_json::to_value(result?).map_err(|err| format!("serialize response failed: {}", err))
 }
@@ -2940,12 +2541,6 @@ fn dispatch_save_general_config(args: &Value) -> Result<Value, String> {
     ))
 }
 
-fn app_handle() -> Result<tauri::AppHandle, String> {
-    crate::get_app_handle()
-        .cloned()
-        .ok_or_else(|| "App runtime is not available".to_string())
-}
-
 fn arg<T: DeserializeOwned>(args: &Value, key: &str) -> Result<T, String> {
     let value = args
         .get(key)
@@ -2989,194 +2584,5 @@ fn opt_nullable_arg<T: DeserializeOwned>(
         Some(value) => serde_json::from_value(value.clone())
             .map(|value| Some(Some(value)))
             .map_err(|err| format!("invalid argument '{}': {}", key, err)),
-    }
-}
-
-async fn read_http_request(stream: &mut TcpStream) -> Result<Option<HttpRequest>, String> {
-    let mut buffer = Vec::new();
-    let mut temp = [0u8; 4096];
-    let header_end = loop {
-        let read = timeout(REQUEST_READ_TIMEOUT, stream.read(&mut temp))
-            .await
-            .map_err(|_| "request read timed out".to_string())?
-            .map_err(|err| err.to_string())?;
-        if read == 0 {
-            if buffer.is_empty() {
-                return Ok(None);
-            }
-            return Err("connection closed before headers completed".to_string());
-        }
-        buffer.extend_from_slice(&temp[..read]);
-        if buffer.len() > MAX_HTTP_REQUEST_BYTES {
-            return Err("request too large".to_string());
-        }
-        if let Some(pos) = find_header_end(&buffer) {
-            break pos;
-        }
-    };
-
-    let header_text =
-        String::from_utf8(buffer[..header_end].to_vec()).map_err(|err| err.to_string())?;
-    let mut lines = header_text.split("\r\n");
-    let request_line = lines
-        .next()
-        .ok_or_else(|| "missing request line".to_string())?;
-    let mut request_parts = request_line.split_whitespace();
-    let method = request_parts.next().unwrap_or("").to_string();
-    let raw_path = request_parts.next().unwrap_or("/");
-    let (path, query) = normalize_request_path(raw_path)?;
-    let mut content_length = 0usize;
-    for line in lines {
-        if let Some((name, value)) = line.split_once(':') {
-            let name = name.trim().to_ascii_lowercase();
-            let value = value.trim().to_string();
-            if name == "content-length" {
-                content_length = value
-                    .parse::<usize>()
-                    .map_err(|_| "invalid content-length".to_string())?;
-            }
-        }
-    }
-
-    if content_length > MAX_HTTP_REQUEST_BYTES {
-        return Err("request body too large".to_string());
-    }
-
-    let body_start = header_end + 4;
-    let mut body = buffer.get(body_start..).unwrap_or_default().to_vec();
-    while body.len() < content_length {
-        let read = timeout(REQUEST_READ_TIMEOUT, stream.read(&mut temp))
-            .await
-            .map_err(|_| "request body read timed out".to_string())?
-            .map_err(|err| err.to_string())?;
-        if read == 0 {
-            return Err("connection closed before body completed".to_string());
-        }
-        body.extend_from_slice(&temp[..read]);
-        if body.len() > MAX_HTTP_REQUEST_BYTES {
-            return Err("request body too large".to_string());
-        }
-    }
-    body.truncate(content_length);
-
-    Ok(Some(HttpRequest {
-        method,
-        path,
-        query,
-        body,
-    }))
-}
-
-fn normalize_request_path(raw_path: &str) -> Result<(String, Option<String>), String> {
-    let url = Url::parse(&format!("http://127.0.0.1{}", raw_path))
-        .map_err(|err| format!("invalid request path: {}", err))?;
-    Ok((url.path().to_string(), url.query().map(str::to_string)))
-}
-
-fn query_param(query: Option<&str>, name: &str) -> Option<String> {
-    let query = query?;
-    url::form_urlencoded::parse(query.as_bytes())
-        .find(|(key, _)| key == name)
-        .map(|(_, value)| value.into_owned())
-}
-
-fn find_header_end(buffer: &[u8]) -> Option<usize> {
-    buffer.windows(4).position(|window| window == b"\r\n\r\n")
-}
-
-async fn write_response(
-    stream: &mut TcpStream,
-    status: u16,
-    reason: &str,
-    content_type: &str,
-    body: &[u8],
-) -> Result<(), String> {
-    let headers = format!(
-        "HTTP/1.1 {} {}\r\ncontent-type: {}\r\ncontent-length: {}\r\ncache-control: no-store\r\nx-content-type-options: nosniff\r\naccess-control-allow-origin: http://127.0.0.1:{}\r\naccess-control-allow-methods: GET,POST,OPTIONS\r\naccess-control-allow-headers: content-type\r\nconnection: close\r\n\r\n",
-        status,
-        reason,
-        content_type,
-        body.len(),
-        get_actual_port().unwrap_or(DEFAULT_WEB_CONSOLE_PORT)
-    );
-    stream
-        .write_all(headers.as_bytes())
-        .await
-        .map_err(|err| err.to_string())?;
-    stream
-        .write_all(body)
-        .await
-        .map_err(|err| err.to_string())?;
-    stream.shutdown().await.map_err(|err| err.to_string())
-}
-
-fn find_frontend_dist() -> Option<PathBuf> {
-    let manifest_dir = Path::new(env!("CARGO_MANIFEST_DIR"));
-    let candidates = [
-        manifest_dir.join("../dist"),
-        std::env::current_dir().ok()?.join("dist"),
-        std::env::current_dir().ok()?.join("../dist"),
-    ];
-
-    candidates
-        .into_iter()
-        .map(|path| normalize_path(&path))
-        .find(|path| path.join(INDEX_HTML).exists())
-}
-
-fn normalize_path(path: &Path) -> PathBuf {
-    path.components().fold(PathBuf::new(), |mut acc, part| {
-        match part {
-            Component::CurDir => {}
-            Component::ParentDir => {
-                acc.pop();
-            }
-            other => acc.push(other.as_os_str()),
-        }
-        acc
-    })
-}
-
-fn resolve_static_path(root: &Path, request_path: &str) -> Result<PathBuf, String> {
-    let path = if request_path == "/" || request_path.is_empty() {
-        INDEX_HTML.to_string()
-    } else {
-        request_path.trim_start_matches('/').to_string()
-    };
-    let decoded =
-        urlencoding::decode(&path).map_err(|err| format!("invalid URL encoding: {}", err))?;
-    let mut result = PathBuf::from(root);
-    for segment in decoded.split('/') {
-        if segment.is_empty() {
-            continue;
-        }
-        if segment == "." || segment == ".." || segment.contains('\\') {
-            return Err("invalid static path".to_string());
-        }
-        result.push(segment);
-    }
-    Ok(result)
-}
-
-fn content_type_for_path(path: &Path) -> &'static str {
-    match path
-        .extension()
-        .and_then(|value| value.to_str())
-        .unwrap_or_default()
-        .to_ascii_lowercase()
-        .as_str()
-    {
-        "html" => "text/html; charset=utf-8",
-        "js" | "mjs" => "text/javascript; charset=utf-8",
-        "css" => "text/css; charset=utf-8",
-        "json" => "application/json; charset=utf-8",
-        "svg" => "image/svg+xml",
-        "png" => "image/png",
-        "jpg" | "jpeg" => "image/jpeg",
-        "webp" => "image/webp",
-        "ico" => "image/x-icon",
-        "woff" => "font/woff",
-        "woff2" => "font/woff2",
-        _ => "application/octet-stream",
     }
 }
